MIM-168 Split oversized daemon and Mac tests
